Move useError back to AddTokens widget from common
