+++ packages/checkout/widgets-lib/src/widgets/add-tokens/components/TokenDrawerMenu.tsx
@@ -26,7 +26,7 @@ import {
   AddTokensActions,
   AddTokensState,
 } from '../context/AddTokensContext';
-import { useError } from '../../../lib/squid/hooks/useError';
+import { useError } from '../hooks/useError';
 import {
   getDefaultTokenImage,
   getTokenImageByAddress,
