+++ packages/checkout/widgets-lib/src/widgets/add-tokens/hooks/useErrorHandler.ts
@@ -2,7 +2,7 @@ import { useContext } from 'react';
 import { AddTokensError, AddTokensErrorTypes } from '../types';
 import { EventTargetContext } from '../../../context/event-target-context/EventTargetContext';
 import { sendAddTokensFailedEvent } from '../AddTokensWidgetEvents';
-import { useError } from '../../../lib/squid/hooks/useError';
+import { useError } from './useError';
 import { AddTokensContext } from '../context/AddTokensContext';
 import { useProvidersContext } from '../../../context/providers-context/ProvidersContext';
